requirements.txt
sentence-transformers>=2.2.0
transformers>=4.30.0
accelerate>=0.20.0
torch>=2.0.0
rank-bm25>=0.2.2
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
beautifulsoup4>=4.12.0
requests>=2.28.0
nltk>=3.8.0
faiss-cpu>=1.7.4
streamlit>=1.25.0
pandas>=2.0.0
matplotlib>=3.7.0
rouge-score>=0.1.2
bert-score>=0.3.13
